Create primary button style
- Refactor today list view
